Correction clear method()
The clear method was deleting the array pointed by data_
(which is unnecessary) and re-writing '\0' in the data[0]
without using a new char[]. It is a source of bug/memory leek because we
were writing directly in the heap.
Now clear() method just replaces the first char of the array
pointed by data_ by '\0'.